Rename `/gpx_export` to `/export`
